add test case for does not send message on creation of SUT and should invoke retrieval message when request cache retrieval. move feed store spy to its own file to share among different tests
