+++ FeedFeature/Feed Cache/FeedStore.swift
@@ -13,4 +13,5 @@ public protocol FeedStore {
 
     func deleteCachedFeed(completion: @escaping DeletionCompletion)
     func insertFeed(items: [LocalFeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+    func retrieveFeed()
 }
