+++ FeedFeatureTests/Feed Cache/LoadFeedUseCaseTests.swift
@@ -1,0 +1,33 @@
+//
+//  LoadFeedUseCaseTests.swift
+//  FeedFeatureTests
+//
+//  Created by Srinivasan Rajendran on 2020-12-25.
+//
+
+import XCTest
+import FeedFeature
+
+class LoadFeedFromCacheUseCaseTests: XCTestCase {
+
+    func test_init_doesNotMessageStoreOnCreation() {
+        let (_, store) = makeSUT()
+        XCTAssertEqual(store.receivedMessages, [])
+    }
+
+    func test_load_shouldRequestCacheRetrieval() {
+        let (sut, store) = makeSUT()
+        sut.load()
+        XCTAssertEqual(store.receivedMessages, [.retrieve])
+    }
+
+    // MARK: Helpers
+
+    private func makeSUT(currentDate: @escaping () -> Date = Date.init ,file: StaticString = #filePath, line: UInt = #line) -> (sut: LocalFeedLoader, store: FeedStoreSpy) {
+        let store = FeedStoreSpy()
+        let sut = LocalFeedLoader(store: store, currentDate: currentDate)
+        trackForMemoryLeaks(sut,file: file, line: line)
+        trackForMemoryLeaks(store,file: file, line: line)
+        return (sut, store)
+    }
+}
